Format Overhaul
input() changed to main() because input is a built-in function not sure how we thought we could get away with that. All functions defined before main is called. This was tested in an online python environment and the input works, but can't call geckodriver, someone needs to test all of this with geckodriver in the directory.